Squashed commit of the following:
    Merge branch 'gerrit_master_24jun18_e758bda' and add Li's updated files

+++ jobs/rocoto/getic_gsd.sh
@@ -59,35 +59,34 @@ echo "PSLOT = $PSLOT"
 echo
 
 ## initialize
-fv3ic_dir=$ICSDIR/$CDATE/$CDUMP
+yyyymmdd=`echo $CDATE | cut -c1-8`
+hh=`echo $CDATE | cut -c9-10`
+yyddd=`date +%y%j -u -d $yyyymmdd`
+fv3ic_dir=$ICSDIR/${CDATE}/${CDUMP}
 
 ## create links in FV3ICS directory
 mkdir -p $fv3ic_dir
 cd $fv3ic_dir
 echo "making link to nemsio files under $fv3ic_dir"
-
-ln -fs $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.sfcanl.nemsio sfcanl.gfs.${CDATE}
-ln -fs $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.atmanl.nemsio siganl.gfs.${CDATE}
-
-
-
-# /5year/NCEPDEV/emc-global/emc.glopara/WCOSS_C/Q1FY19/prfv3rt1/201804110/gdas.tar 
-#      ./gdas.20180411/00/gdas.t00z.atmanl.nemsio
-#      ./gdas.20180411/00/gdas.t00z.sfcanl.nemsio
-#
-
-# check for correct number of parameters
-
-#if [ $# -lt 2 ]; then
-#  echo "   Usage:  $0  YYYYMMDD HH"
-#  exit 1
-#fi
-
-#yyyymmdd=$1
-#hh=$2
-
+if [[ -f $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.sfcanl.nemsio ]]
+then
+  ln -fs $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.sfcanl.nemsio sfcanl.gfs.${CDATE}
+  ln -fs $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.atmanl.nemsio siganl.gfs.${CDATE}
+  if [[ -f $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.nstanl.nemsio ]] 
+  then 
+    ln -fs $PUBDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.nstanl.nemsio nstanl.gfs.${CDATE}
+  fi
+else 
+  if [[ -f $ARCDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.sfcanl.nemsio ]]
+  then
+    ln -fs $ARCDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.sfcanl.nemsio sfcanl.gfs.${CDATE}
+    ln -fs $ARCDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.atmanl.nemsio siganl.gfs.${CDATE}
+    if [[ -f $ARCDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.nstanl.nemsio ]] 
+    then
+      ln -fs $ARCDIR/${yyddd}${hh}00.${CDUMP}.t${hh}z.nstanl.nemsio nstanl.gfs.${CDATE}
+    fi 
+  fi
 echo "YYYYMMDDHH:  ${yyyymmdd}${hh}"
-
 gdasfile=$GDASDIR/${yyyymmdd}${hh}/gdas.tar
 hsi -q list $gdasfile
 status=$?
